+++ ios/Classes/SmileIDBiometricKYC.swift
@@ -3,7 +3,7 @@ import UIKit
 import SmileID
 import SwiftUI
 
-class SmileIDBiometricKYC : NSObject, FlutterPlatformView, BiometricKycResultDelegate,SmileIDFileUtilsProtocol {
+class SmileIDBiometricKYC : NSObject, FlutterPlatformView, BiometricKycResultDelegate, SmileIDFileUtilsProtocol {
     var fileManager: FileManager = Foundation.FileManager.default
     private var _view: UIView
     private var _channel: FlutterMethodChannel